+++ src/levenshtein_myers_16x8.cpp
@@ -6,7 +6,7 @@
 static const uint16x8_t NULL_V_16 = vdupq_n_u16(0);
 static const uint16x8_t ONE_V_16 = vdupq_n_u16(1);
 
-std::array<uint32_t, 8> levenshtein_myers_16x8(const Myers16x8Input &input) {
+std::array<uint16_t, 8> levenshtein_myers_16x8(const Myers16x8Input &input) {
   uint16_t bm[ALPHABET_LEN] = {0}; // Bitmap for each letter in the alphabet
 
   const char *q_wrd = input.q_wrd;
